📦 refactor: rename dtos/<domain>.dto.ts to dtos/<domain>-dto.ts
- Align file naming with the hyphenated convention used elsewhere in the project
