+++ src/dtos/index.ts
@@ -1,5 +1,5 @@
-export * from './activation.dto'
-export * from './migration.dto'
-export * from './session.dto'
-export * from './status.dto'
-export * from './user.dto'
+export * from './activation-dto'
+export * from './migration-dto'
+export * from './session-dto'
+export * from './status-dto'
+export * from './user-dto'
